+++ src/components/common/blocks/CommonControlBlock.vue
@@ -5,6 +5,7 @@
         class="control-block"
         :class="[
             `control-block--location-${ location }`,
+            `control-block--center-by-${ centerBy }`,
             {
                 'control-block--disabled': disabled,
                 'control-block--has-max-height': maxHeight,
@@ -35,10 +36,18 @@ import type { TooltipLocation } from '~/components/common/basic/CommonTooltip.vu
 import CloseIcon from '@/assets/icons/basic/close.svg?component';
 import { useClickOutside } from '~/composables/click-outside';
 
-defineProps({
+const props = defineProps({
     location: {
         type: String as PropType<TooltipLocation>,
         default: 'top',
+    },
+    centerBy: {
+        type: String as PropType<'start' | 'center' | 'middle' | 'end'>,
+        default: 'center',
+    },
+    centerByOffset: {
+        type: String,
+        default: '0',
     },
     width: {
         type: String,
@@ -76,6 +85,7 @@ const model = defineModel({
 useClickOutside({
     element: block,
     callback: () => {
+        if (!props.closeByClickOutside) return;
         model.value = false;
     },
 });
@@ -83,37 +93,52 @@ useClickOutside({
 
 <style scoped lang="scss">
 .control-block {
+    --transform: calc(-50% + v-bind(centerByOffset));
     position: absolute;
+    z-index: 5;
 
     width: v-bind(width);
     max-width: v-bind(maxWidth);
     padding: 16px;
 
     background: $neutral1000;
+    border: 1px solid varToRgba('neutral150', 0.15);
     border-radius: 8px;
+
+    &--center-by-start {
+        --transform: calc(0% + v-bind(centerByOffset));
+    }
+
+    &--center-by-end {
+        --transform: calc(50% + v-bind(centerByOffset));
+    }
+
+    &--center-by-middle {
+        --transform: calc(25% + v-bind(centerByOffset));
+    }
 
     &--location-top {
         bottom: calc(100% + v-bind(gapFromParent));
         left: auto;
-        transform: translateX(-50%);
+        transform: translateX(var(--transform));
     }
 
     &--location-bottom {
         top: calc(100% + v-bind(gapFromParent));
         left: auto;
-        transform: translateX(-50%);
+        transform: translateX(var(--transform));
     }
 
     &--location-left {
         top: auto;
         right: calc(100% + v-bind(gapFromParent));
-        transform: translateY(-50%);
+        transform: translateY(var(--transform));
     }
 
     &--location-right {
         top: auto;
         left: calc(100% + v-bind(gapFromParent));
-        transform: translateY(-50%);
+        transform: translateY(var(--transform));
     }
 
     &--has-max-height {
@@ -140,6 +165,13 @@ useClickOutside({
             cursor: pointer;
             width: 14px;
             min-width: 14px;
+            transition: 0.3s;
+
+            @include hover {
+                &:hover {
+                    color: $error500;
+                }
+            }
         }
     }